Extract LoggerTrait from DriverBase
